Refactored CoordinateAreas and order of functions to improve understanding
